requirements.txt
# Copyright © 2025 Wenze Wei
#
# This file is part of Pisces.
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU Affero General Public License as published
# by the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
# GNU Affero General Public License for more details.
#
# You should have received a copy of the GNU Affero General Public License
# along with this program. If not, see <https://www.gnu.org/licenses/>.

torch>=2.1.0
transformers>=4.40.0
tokenizers>=0.19.0
datasets>=2.18.0
accelerate>=0.28.0
huggingface-hub>=0.30.2
einops>=0.7.0
opencv-python>=4.8.0
pillow>=11.2.1
timm>=0.9.0
PyMuPDF>=1.23.0